ReadMe update and QATv2 Refactoring
